+++ postgres.go
@@ -145,10 +145,11 @@ func (store *PostgresStore) SaveVisitorStats(tx *sqlx.Tx, entity *VisitorStats) 
 	}
 
 	existing := new(VisitorStats)
-	err := tx.Get(existing, `SELECT id, visitors, sessions, bounces, platform_desktop, platform_mobile, platform_unknown FROM "visitor_stats"
+	err := tx.Get(existing, `SELECT id, visitors, sessions, bounces, platform_desktop, platform_mobile, platform_unknown
+		FROM "visitor_stats"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" = $2
-		AND LOWER("path") = LOWER($3)`, entity.TenantID, entity.Day, entity.Path)
+		AND (LOWER("path") = LOWER($3) OR $3 IS NULL AND "path" IS NULL)`, entity.TenantID, entity.Day, entity.Path)
 
 	if err == nil {
 		existing.Visitors += entity.Visitors
@@ -192,8 +193,7 @@ func (store *PostgresStore) SaveVisitorTimeStats(tx *sqlx.Tx, entity *VisitorTim
 	err := tx.Get(existing, `SELECT id, visitors FROM "visitor_time_stats"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" = $2
-		AND LOWER("path") = LOWER($3)
-		AND "hour" = $4`, entity.TenantID, entity.Day, entity.Path, entity.Hour)
+		AND "hour" = $3`, entity.TenantID, entity.Day, entity.Hour)
 
 	if err == nil {
 		existing.Visitors += entity.Visitors
@@ -204,7 +204,7 @@ func (store *PostgresStore) SaveVisitorTimeStats(tx *sqlx.Tx, entity *VisitorTim
 			return err
 		}
 	} else {
-		rows, err := tx.NamedQuery(`INSERT INTO "visitor_time_stats" ("tenant_id", "day", "path", "hour", "visitors") VALUES (:tenant_id, :day, :path, :hour, :visitors)`, entity)
+		rows, err := tx.NamedQuery(`INSERT INTO "visitor_time_stats" ("tenant_id", "day", "hour", "visitors") VALUES (:tenant_id, :day, :hour, :visitors)`, entity)
 
 		if err != nil {
 			return err
@@ -227,7 +227,7 @@ func (store *PostgresStore) SaveLanguageStats(tx *sqlx.Tx, entity *LanguageStats
 	err := tx.Get(existing, `SELECT id, visitors FROM "language_stats"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" = $2
-		AND LOWER("path") = LOWER($3)
+		AND (LOWER("path") = LOWER($3) OR $3 IS NULL AND "path" IS NULL)
 		AND LOWER("language") = LOWER($4)`, entity.TenantID, entity.Day, entity.Path, entity.Language)
 
 	if err := store.createUpdateEntity(tx, entity, existing, err == nil,
@@ -250,7 +250,7 @@ func (store *PostgresStore) SaveReferrerStats(tx *sqlx.Tx, entity *ReferrerStats
 	err := tx.Get(existing, `SELECT id, visitors FROM "referrer_stats"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" = $2
-		AND LOWER("path") = LOWER($3)
+		AND (LOWER("path") = LOWER($3) OR $3 IS NULL AND "path" IS NULL)
 		AND LOWER("referrer") = LOWER($4)`, entity.TenantID, entity.Day, entity.Path, entity.Referrer)
 
 	if err := store.createUpdateEntity(tx, entity, existing, err == nil,
@@ -273,7 +273,7 @@ func (store *PostgresStore) SaveOSStats(tx *sqlx.Tx, entity *OSStats) error {
 	err := tx.Get(existing, `SELECT id, visitors FROM "os_stats"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" = $2
-		AND LOWER("path") = LOWER($3)
+		AND (LOWER("path") = LOWER($3) OR $3 IS NULL AND "path" IS NULL)
 		AND "os" = $4
 		AND "os_version" = $5`, entity.TenantID, entity.Day, entity.Path, entity.OS, entity.OSVersion)
 
@@ -297,7 +297,7 @@ func (store *PostgresStore) SaveBrowserStats(tx *sqlx.Tx, entity *BrowserStats) 
 	err := tx.Get(existing, `SELECT id, visitors FROM "browser_stats"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" = $2
-		AND LOWER("path") = LOWER($3)
+		AND (LOWER("path") = LOWER($3) OR $3 IS NULL AND "path" IS NULL)
 		AND "browser" = $4
 		AND "browser_version" = $5`, entity.TenantID, entity.Day, entity.Path, entity.Browser, entity.BrowserVersion)
 
@@ -417,6 +417,7 @@ func (store *PostgresStore) Paths(tenantID sql.NullInt64, from, to time.Time) ([
 			WHERE ($1::bigint IS NULL OR tenant_id = $1)
 			AND "day" >= $2::date
 			AND "day" <= $3::date
+    		AND "path" IS NOT NULL
 		) AS results
 		ORDER BY "path" ASC`
 	var paths []string
@@ -513,7 +514,7 @@ func (store *PostgresStore) CountVisitorsByPath(tx *sqlx.Tx, tenantID sql.NullIn
 }
 
 // CountVisitorsByPathAndHour implements the Store interface.
-func (store *PostgresStore) CountVisitorsByPathAndHour(tx *sqlx.Tx, tenantID sql.NullInt64, day time.Time, path string) ([]VisitorTimeStats, error) {
+func (store *PostgresStore) CountVisitorsByHour(tx *sqlx.Tx, tenantID sql.NullInt64, day time.Time) ([]VisitorTimeStats, error) {
 	if tx == nil {
 		tx = store.NewTx()
 		defer store.Commit(tx)
@@ -521,21 +522,18 @@ func (store *PostgresStore) CountVisitorsByPathAndHour(tx *sqlx.Tx, tenantID sql
 
 	query := `SELECT $1::bigint AS "tenant_id",
 		$2::date AS "day",
-		$3::varchar AS "path",
 		EXTRACT(HOUR FROM "day_and_hour") "hour",
 		(
 			SELECT count(DISTINCT "fingerprint") FROM "hit"
 			WHERE ($1::bigint IS NULL OR tenant_id = $1)
 			AND "time" >= "day_and_hour"
 			AND "time" < "day_and_hour" + INTERVAL '1 hour'
-			AND LOWER("path") = LOWER($3)
 		) "visitors",
        (
 			SELECT count(DISTINCT("fingerprint", "session")) FROM "hit"
 			WHERE ($1::bigint IS NULL OR tenant_id = $1)
 			AND "time" >= "day_and_hour"
 			AND "time" < "day_and_hour" + INTERVAL '1 hour'
-			AND LOWER("path") = LOWER($3)
 		) "sessions"
 		FROM (
 			SELECT * FROM generate_series(
@@ -546,7 +544,7 @@ func (store *PostgresStore) CountVisitorsByPathAndHour(tx *sqlx.Tx, tenantID sql
 		) AS hours`
 	var visitors []VisitorTimeStats
 
-	if err := tx.Select(&visitors, query, tenantID, day, path); err != nil {
+	if err := tx.Select(&visitors, query, tenantID, day); err != nil {
 		return nil, err
 	}
 
@@ -661,7 +659,7 @@ func (store *PostgresStore) CountVisitorsByLanguage(tx *sqlx.Tx, tenantID sql.Nu
 		defer store.Commit(tx)
 	}
 
-	query := `SELECT "language", count(DISTINCT fingerprint) "visitors"
+	query := `SELECT "language", count(DISTINCT fingerprint) "visitors", $2::date "day"
 		FROM "hit"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND date("time") = $2::date
@@ -682,7 +680,7 @@ func (store *PostgresStore) CountVisitorsByReferrer(tx *sqlx.Tx, tenantID sql.Nu
 		defer store.Commit(tx)
 	}
 
-	query := `SELECT "referrer", count(DISTINCT fingerprint) "visitors"
+	query := `SELECT "referrer", count(DISTINCT fingerprint) "visitors", $2::date "day"
 		FROM "hit"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND date("time") = $2::date
@@ -703,7 +701,7 @@ func (store *PostgresStore) CountVisitorsByOS(tx *sqlx.Tx, tenantID sql.NullInt6
 		defer store.Commit(tx)
 	}
 
-	query := `SELECT "os", count(DISTINCT fingerprint) "visitors"
+	query := `SELECT "os", count(DISTINCT fingerprint) "visitors", $2::date "day"
 		FROM "hit"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND date("time") = $2::date
@@ -724,7 +722,7 @@ func (store *PostgresStore) CountVisitorsByBrowser(tx *sqlx.Tx, tenantID sql.Nul
 		defer store.Commit(tx)
 	}
 
-	query := `SELECT "browser", count(DISTINCT fingerprint) "visitors"
+	query := `SELECT "browser", count(DISTINCT fingerprint) "visitors", $2::date "day"
 		FROM "hit"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND date("time") = $2::date
@@ -900,7 +898,7 @@ func (store *PostgresStore) Visitors(tenantID sql.NullInt64, from, to time.Time)
 				INTERVAL '1 day'
 			) "d"
 		) AS date_series
-		LEFT JOIN "visitor_stats" ON ($1::bigint IS NULL OR tenant_id = $1) AND "visitor_stats"."day" = "d"
+		LEFT JOIN "visitor_stats" ON ($1::bigint IS NULL OR tenant_id = $1) AND "visitor_stats"."day" = "d" AND "visitor_stats"."path" IS NULL
 		GROUP BY "d"
 		ORDER BY "d" ASC`
 	var visitors []Stats
@@ -950,6 +948,7 @@ func (store *PostgresStore) VisitorLanguages(tenantID sql.NullInt64, from, to ti
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" >= $2::date
 		AND "day" <= $3::date
+		AND "path" IS NULL
 		GROUP BY "language"
 		ORDER BY "visitors" DESC`
 	var visitors []LanguageStats
@@ -968,6 +967,7 @@ func (store *PostgresStore) VisitorReferrer(tenantID sql.NullInt64, from, to tim
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" >= $2::date
 		AND "day" <= $3::date
+		AND "path" IS NULL
 		GROUP BY "referrer"
 		ORDER BY "visitors" DESC`
 	var visitors []ReferrerStats
@@ -986,6 +986,7 @@ func (store *PostgresStore) VisitorOS(tenantID sql.NullInt64, from, to time.Time
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" >= $2::date
 		AND "day" <= $3::date
+		AND "path" IS NULL
 		GROUP BY "os"
 		ORDER BY "visitors" DESC`
 	var visitors []OSStats
@@ -1004,6 +1005,7 @@ func (store *PostgresStore) VisitorBrowser(tenantID sql.NullInt64, from, to time
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" >= $2::date
 		AND "day" <= $3::date
+		AND "path" IS NULL
 		GROUP BY "browser"
 		ORDER BY "visitors" DESC`
 	var visitors []BrowserStats
@@ -1023,7 +1025,8 @@ func (store *PostgresStore) VisitorPlatform(tenantID sql.NullInt64, from, to tim
 		FROM "visitor_stats"
 		WHERE ($1::bigint IS NULL OR tenant_id = $1)
 		AND "day" >= $2::date
-		AND "day" <= $3::date`
+		AND "day" <= $3::date
+		AND "path" IS NULL`
 	visitors := new(VisitorStats)
 
 	if err := store.DB.Get(visitors, query, tenantID, from, to); err != nil && err != sql.ErrNoRows {
@@ -1295,6 +1298,8 @@ func (store *PostgresStore) VisitorsSum(tenantID sql.NullInt64, from, to time.Ti
 	if path != "" {
 		args = append(args, path)
 		query += `AND LOWER("path") = LOWER($4) `
+	} else {
+		query += `AND "path" IS NULL `
 	}
 
 	visitors := new(Stats)