+++ processor_test.go
@@ -16,78 +16,124 @@ func TestProcessor_ProcessTenant(t *testing.T) {
 }
 
 func TestProcessor_ProcessSessions(t *testing.T) {
-	for _, store := range testStorageBackends() {
-		cleanupDB(t)
+	store := NewPostgresStore(postgresDB, nil)
+	cleanupDB(t)
 
-		// create hits for two visitors and three sessions
-		now := time.Now()
-		createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 9, 7, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
-		createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 9, 7, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
-		createHit(t, store, 0, "fp2", "/", "en", "", "", day(2020, 9, 7, 5), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
-		createHit(t, store, 0, "fp2", "/", "en", "", "", day(2020, 9, 7, 5), now.Add(time.Second*1), OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
-		processor := NewProcessor(store)
+	// create hits for two visitors and three sessions
+	now := time.Now()
+	createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 9, 7, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 9, 7, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp2", "/", "en", "", "", day(2020, 9, 7, 5), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp2", "/", "en", "", "", day(2020, 9, 7, 5), now.Add(time.Second*1), OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	processor := NewProcessor(store)
 
-		if err := processor.Process(); err != nil {
-			t.Fatalf("Data must have been processed, but was: %v", err)
-		}
+	if err := processor.Process(); err != nil {
+		t.Fatalf("Data must have been processed, but was: %v", err)
+	}
 
-		checkHits(t, 0)
-		db := sqlx.NewDb(postgresDB, "postgres")
-		var visitorStats []VisitorStats
-		var timeStats []VisitorTimeStats
+	checkHits(t, 0)
+	db := sqlx.NewDb(postgresDB, "postgres")
+	var visitorStats []VisitorStats
+	var timeStats []VisitorTimeStats
 
-		if err := db.Select(&visitorStats, `SELECT * FROM "visitor_stats" ORDER BY "day", "path"`); err != nil {
-			t.Fatal(err)
-		}
+	if err := db.Select(&visitorStats, `SELECT * FROM "visitor_stats" ORDER BY "day", "path"`); err != nil {
+		t.Fatal(err)
+	}
 
-		if err := db.Select(&timeStats, `SELECT * FROM "visitor_time_stats" ORDER BY "day", "path"`); err != nil {
-			t.Fatal(err)
-		}
+	if err := db.Select(&timeStats, `SELECT * FROM "visitor_time_stats" ORDER BY "day", "hour"`); err != nil {
+		t.Fatal(err)
+	}
 
-		if len(visitorStats) != 1 {
-			t.Fatalf("One visitor stats must have been created, but was: %v", len(visitorStats))
-		}
+	if len(visitorStats) != 2 {
+		t.Fatalf("Two visitor stats must have been created, but was: %v", len(visitorStats))
+	}
 
-		if len(timeStats) != 24 {
-			t.Fatalf("24 visitor time stats must have been created, but was: %v", len(visitorStats))
-		}
+	if len(timeStats) != 24 {
+		t.Fatalf("24 visitor time stats must have been created, but was: %v", len(visitorStats))
+	}
 
-		if visitorStats[0].Visitors != 2 ||
-			visitorStats[0].Sessions != 3 {
-			t.Fatalf("Visitor stats must have two visitors and three sessions, but was: %v %v", visitorStats[0].Visitors, visitorStats[0].Sessions)
-		}
+	if visitorStats[0].Visitors != 2 || visitorStats[0].Sessions != 3 {
+		t.Fatalf("Visitor stats must have two visitors and three sessions, but was: %v %v", visitorStats[0].Visitors, visitorStats[0].Sessions)
+	}
 
-		if timeStats[4].Visitors != 1 || timeStats[5].Visitors != 1 {
-			t.Fatalf("Visitor time stats must have two visitors, but was: %v", timeStats)
-		}
+	if timeStats[4].Visitors != 1 || timeStats[5].Visitors != 1 {
+		t.Fatalf("Visitor time stats must have two visitors, but was: %v %v", timeStats[4].Visitors, timeStats[5].Visitors)
+	}
+}
+
+func TestProcessor_ProcessPaths(t *testing.T) {
+	store := NewPostgresStore(postgresDB, nil)
+	cleanupDB(t)
+	now := time.Now()
+	createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 12, 27, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 12, 27, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 12, 27, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp1", "/", "en", "", "", day(2020, 12, 27, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp1", "/path", "en", "", "", day(2020, 12, 27, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp1", "/path", "en", "", "", day(2020, 12, 27, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp1", "/path", "en", "", "", day(2020, 12, 27, 4), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp2", "/path", "en", "", "", day(2020, 12, 27, 5), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp2", "/path", "en", "", "", day(2020, 12, 27, 5), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	createHit(t, store, 0, "fp2", "/path", "en", "", "", day(2020, 12, 27, 5), now, OSWindows, "10", BrowserChrome, "84.0", "", true, false, 0, 0)
+	processor := NewProcessor(store)
+
+	if err := processor.Process(); err != nil {
+		t.Fatalf("Data must have been processed, but was: %v", err)
+	}
+
+	analyzer := NewAnalyzer(store, nil)
+	pageVisitors, err := analyzer.PageVisitors(&Filter{
+		From: day(2020, 12, 25, 0),
+		To:   day(2020, 12, 28, 0),
+	})
+
+	if err != nil {
+		t.Fatalf("Page visitors must have been returned, but was: %v", err)
+	}
+
+	if len(pageVisitors) != 2 ||
+		pageVisitors[0].Path != "/" ||
+		sumUpVisitors(pageVisitors[0].Stats) != 1 ||
+		pageVisitors[1].Path != "/path" ||
+		sumUpVisitors(pageVisitors[1].Stats) != 2 {
+		t.Fatalf("Page visitor statistics not as expected: %v", pageVisitors)
+	}
+
+	visitors, err := analyzer.Visitors(nil)
+
+	if err != nil {
+		t.Fatalf("Visitors must have been returned, but was: %v", err)
+	}
+
+	if sumUpVisitors(visitors) != 2 {
+		t.Fatalf("Visitor statistics not as expected: %v", visitors)
 	}
 }
 
 func testProcess(t *testing.T, tenantID int64) {
-	for _, store := range testStorageBackends() {
-		createTestdata(t, store, tenantID)
-		processor := NewProcessor(store)
+	store := NewPostgresStore(postgresDB, nil)
+	createTestdata(t, store, tenantID)
+	processor := NewProcessor(store)
 
-		if tenantID == 0 {
-			if err := processor.Process(); err != nil {
-				t.Fatalf("Data must have been processed, but was: %v", err)
-			}
-		} else {
-			if err := processor.ProcessTenant(NewTenantID(tenantID)); err != nil {
-				t.Fatalf("Data must have been processed, but was: %v", err)
-			}
+	if tenantID == 0 {
+		if err := processor.Process(); err != nil {
+			t.Fatalf("Data must have been processed, but was: %v", err)
 		}
-
-		checkHits(t, tenantID)
-		checkVisitorStats(t, tenantID)
-		checkVisitorTimeStats(t, tenantID)
-		checkLanguageStats(t, tenantID)
-		checkReferrerStats(t, tenantID)
-		checkOSStats(t, tenantID)
-		checkBrowserStats(t, tenantID)
-		checkScreenStats(t, tenantID)
-		checkCountryStats(t, tenantID)
+	} else {
+		if err := processor.ProcessTenant(NewTenantID(tenantID)); err != nil {
+			t.Fatalf("Data must have been processed, but was: %v", err)
+		}
 	}
+
+	checkHits(t, tenantID)
+	checkVisitorStats(t, tenantID)
+	checkVisitorTimeStats(t, tenantID)
+	checkLanguageStats(t, tenantID)
+	checkReferrerStats(t, tenantID)
+	checkOSStats(t, tenantID)
+	checkBrowserStats(t, tenantID)
+	checkScreenStats(t, tenantID)
+	checkCountryStats(t, tenantID)
 }
 
 func checkHits(t *testing.T, tenantID int64) {
@@ -114,11 +160,11 @@ func checkVisitorStats(t *testing.T, tenantID int64) {
 	var stats []VisitorStats
 
 	if tenantID != 0 {
-		if err := db.Select(&stats, `SELECT * FROM "visitor_stats" WHERE tenant_id = $1 ORDER BY "day", "path"`, tenantID); err != nil {
+		if err := db.Select(&stats, `SELECT * FROM "visitor_stats" WHERE tenant_id = $1 AND "path" IS NOT NULL ORDER BY "day", "path"`, tenantID); err != nil {
 			t.Fatal(err)
 		}
 	} else {
-		if err := db.Select(&stats, `SELECT * FROM "visitor_stats" ORDER BY "day", "path"`); err != nil {
+		if err := db.Select(&stats, `SELECT * FROM "visitor_stats" WHERE "path" IS NOT NULL ORDER BY "day", "path"`); err != nil {
 			t.Fatal(err)
 		}
 	}
@@ -127,10 +173,10 @@ func checkVisitorStats(t *testing.T, tenantID int64) {
 		t.Fatalf("Four stats must have been created, but was: %v", len(stats))
 	}
 
-	if stats[0].Path != "/" || stats[0].Visitors != 2 || stats[0].PlatformDesktop != 2 || stats[0].PlatformMobile != 0 || stats[0].PlatformUnknown != 0 || stats[0].Bounces != 2 ||
-		stats[1].Path != "/page" || stats[1].Visitors != 1 || stats[1].PlatformDesktop != 1 || stats[1].PlatformMobile != 0 || stats[1].PlatformUnknown != 0 || stats[1].Bounces != 1 ||
-		stats[2].Path != "/" || stats[2].Visitors != 2 || stats[2].PlatformDesktop != 1 || stats[2].PlatformMobile != 0 || stats[2].PlatformUnknown != 1 || stats[2].Bounces != 2 ||
-		stats[3].Path != "/different-page" || stats[3].Visitors != 1 || stats[3].PlatformDesktop != 0 || stats[3].PlatformMobile != 1 || stats[3].PlatformUnknown != 0 || stats[3].Bounces != 1 {
+	if stats[0].Path.String != "/" || stats[0].Visitors != 2 || stats[0].PlatformDesktop != 2 || stats[0].PlatformMobile != 0 || stats[0].PlatformUnknown != 0 || stats[0].Bounces != 2 ||
+		stats[1].Path.String != "/page" || stats[1].Visitors != 1 || stats[1].PlatformDesktop != 1 || stats[1].PlatformMobile != 0 || stats[1].PlatformUnknown != 0 || stats[1].Bounces != 1 ||
+		stats[2].Path.String != "/" || stats[2].Visitors != 2 || stats[2].PlatformDesktop != 1 || stats[2].PlatformMobile != 0 || stats[2].PlatformUnknown != 1 || stats[2].Bounces != 2 ||
+		stats[3].Path.String != "/different-page" || stats[3].Visitors != 1 || stats[3].PlatformDesktop != 0 || stats[3].PlatformMobile != 1 || stats[3].PlatformUnknown != 0 || stats[3].Bounces != 1 {
 		t.Fatalf("Stats not as expected: %v", stats)
 	}
 }
@@ -140,23 +186,23 @@ func checkVisitorTimeStats(t *testing.T, tenantID int64) {
 	var stats []VisitorTimeStats
 
 	if tenantID != 0 {
-		if err := db.Select(&stats, `SELECT * FROM "visitor_time_stats" WHERE tenant_id = $1 ORDER BY "day", "path", "hour"`, tenantID); err != nil {
+		if err := db.Select(&stats, `SELECT * FROM "visitor_time_stats" WHERE tenant_id = $1 ORDER BY "day", "hour"`, tenantID); err != nil {
 			t.Fatal(err)
 		}
 	} else {
-		if err := db.Select(&stats, `SELECT * FROM "visitor_time_stats" ORDER BY "day", "path", "hour"`); err != nil {
+		if err := db.Select(&stats, `SELECT * FROM "visitor_time_stats" ORDER BY "day", "hour"`); err != nil {
 			t.Fatal(err)
 		}
 	}
 
-	if len(stats) != 96 {
-		t.Fatalf("96 stats must have been created, but was: %v", len(stats))
+	if len(stats) != 48 {
+		t.Fatalf("48 stats must have been created, but was: %v", len(stats))
 	}
 
-	if stats[7].Path != "/" || stats[7].Visitors != 2 || stats[7].Hour != 7 ||
-		stats[32].Path != "/page" || stats[32].Visitors != 1 || stats[32].Hour != 8 ||
-		stats[57].Path != "/" || stats[57].Visitors != 2 || stats[57].Hour != 9 ||
-		stats[82].Path != "/different-page" || stats[82].Visitors != 1 || stats[82].Hour != 10 {
+	if stats[7].Visitors != 2 || stats[7].Hour != 7 ||
+		stats[8].Visitors != 1 || stats[8].Hour != 8 ||
+		stats[24+9].Visitors != 2 || stats[24+9].Hour != 9 ||
+		stats[24+10].Visitors != 1 || stats[24+10].Hour != 10 {
 		t.Fatalf("Stats not as expected: %v", stats)
 	}
 }
@@ -175,14 +221,18 @@ func checkLanguageStats(t *testing.T, tenantID int64) {
 		}
 	}
 
-	if len(stats) != 4 {
+	if len(stats) != 8 {
 		t.Fatalf("Four stats must have been created, but was: %v", len(stats))
 	}
 
-	if stats[0].Path != "/" || stats[0].Visitors != 2 || stats[0].Language.String != "en" ||
-		stats[1].Path != "/page" || stats[1].Visitors != 1 || stats[1].Language.String != "de" ||
-		stats[2].Path != "/" || stats[2].Visitors != 2 || stats[2].Language.String != "en" ||
-		stats[3].Path != "/different-page" || stats[3].Visitors != 1 || stats[3].Language.String != "jp" {
+	if stats[0].Path.String != "/" || stats[0].Visitors != 2 || stats[0].Language.String != "en" ||
+		stats[1].Path.String != "/page" || stats[1].Visitors != 1 || stats[1].Language.String != "de" ||
+		stats[2].Path.Valid || stats[2].Visitors != 1 || stats[2].Language.String != "de" ||
+		stats[3].Path.Valid || stats[3].Visitors != 2 || stats[3].Language.String != "en" ||
+		stats[4].Path.String != "/" || stats[4].Visitors != 2 || stats[4].Language.String != "en" ||
+		stats[5].Path.String != "/different-page" || stats[5].Visitors != 1 || stats[5].Language.String != "jp" ||
+		stats[6].Path.Valid || stats[6].Visitors != 2 || stats[6].Language.String != "en" ||
+		stats[7].Path.Valid || stats[7].Visitors != 1 || stats[7].Language.String != "jp" {
 		t.Fatalf("Stats not as expected: %v", stats)
 	}
 }
@@ -201,14 +251,17 @@ func checkReferrerStats(t *testing.T, tenantID int64) {
 		}
 	}
 
-	if len(stats) != 4 {
-		t.Fatalf("Four stats must have been created, but was: %v", len(stats))
+	if len(stats) != 7 {
+		t.Fatalf("Seven stats must have been created, but was: %v", len(stats))
 	}
 
-	if stats[0].Path != "/" || stats[0].Visitors != 2 || stats[0].Referrer.String != "ref1" ||
-		stats[1].Path != "/page" || stats[1].Visitors != 1 || stats[1].Referrer.String != "ref1" ||
-		stats[2].Path != "/" || stats[2].Visitors != 2 || stats[2].Referrer.String != "ref2" ||
-		stats[3].Path != "/different-page" || stats[3].Visitors != 1 || stats[3].Referrer.String != "ref3" {
+	if stats[0].Path.String != "/" || stats[0].Visitors != 2 || stats[0].Referrer.String != "ref1" ||
+		stats[1].Path.String != "/page" || stats[1].Visitors != 1 || stats[1].Referrer.String != "ref1" ||
+		stats[2].Path.Valid || stats[2].Visitors != 3 || stats[2].Referrer.String != "ref1" ||
+		stats[3].Path.String != "/" || stats[3].Visitors != 2 || stats[3].Referrer.String != "ref2" ||
+		stats[4].Path.String != "/different-page" || stats[4].Visitors != 1 || stats[4].Referrer.String != "ref3" ||
+		stats[5].Path.Valid || stats[5].Visitors != 2 || stats[5].Referrer.String != "ref2" ||
+		stats[6].Path.Valid || stats[6].Visitors != 1 || stats[6].Referrer.String != "ref3" {
 		t.Fatalf("Stats not as expected: %v", stats)
 	}
 }
@@ -227,15 +280,20 @@ func checkOSStats(t *testing.T, tenantID int64) {
 		}
 	}
 
-	if len(stats) != 5 {
-		t.Fatalf("Five stats must have been created, but was: %v", len(stats))
+	if len(stats) != 10 {
+		t.Fatalf("Ten stats must have been created, but was: %v", len(stats))
 	}
 
-	if stats[0].Path != "/" || stats[0].Visitors != 2 || stats[0].OS.String != OSWindows || stats[0].OSVersion.String != "10" ||
-		stats[1].Path != "/page" || stats[1].Visitors != 1 || stats[1].OS.String != OSMac || stats[1].OSVersion.String != "10.15.3" ||
-		stats[2].Path != "/" || stats[2].Visitors != 1 || stats[2].OS.String != OSLinux || stats[2].OSVersion.String != "" ||
-		stats[3].Path != "/" || stats[3].Visitors != 1 || stats[3].OS.String != OSWindows || stats[3].OSVersion.String != "10" ||
-		stats[4].Path != "/different-page" || stats[4].Visitors != 1 || stats[4].OS.String != OSAndroid || stats[4].OSVersion.String != "8.0" {
+	if stats[0].Path.String != "/" || stats[0].Visitors != 2 || stats[0].OS.String != OSWindows || stats[0].OSVersion.String != "10" ||
+		stats[1].Path.String != "/page" || stats[1].Visitors != 1 || stats[1].OS.String != OSMac || stats[1].OSVersion.String != "10.15.3" ||
+		stats[2].Path.Valid || stats[2].Visitors != 1 || stats[2].OS.String != OSMac || stats[2].OSVersion.Valid ||
+		stats[3].Path.Valid || stats[3].Visitors != 2 || stats[3].OS.String != OSWindows || stats[3].OSVersion.Valid ||
+		stats[4].Path.String != "/" || stats[4].Visitors != 1 || stats[4].OS.String != OSLinux || stats[4].OSVersion.String != "" ||
+		stats[5].Path.String != "/" || stats[5].Visitors != 1 || stats[5].OS.String != OSWindows || stats[5].OSVersion.String != "10" ||
+		stats[6].Path.String != "/different-page" || stats[6].Visitors != 1 || stats[6].OS.String != OSAndroid || stats[6].OSVersion.String != "8.0" ||
+		stats[7].Path.Valid || stats[7].Visitors != 1 || stats[7].OS.String != OSAndroid || stats[7].OSVersion.Valid ||
+		stats[8].Path.Valid || stats[8].Visitors != 1 || stats[8].OS.String != OSLinux || stats[8].OSVersion.Valid ||
+		stats[9].Path.Valid || stats[9].Visitors != 1 || stats[9].OS.String != OSWindows || stats[9].OSVersion.Valid {
 		t.Fatalf("Stats not as expected: %v", stats)
 	}
 }
@@ -254,15 +312,18 @@ func checkBrowserStats(t *testing.T, tenantID int64) {
 		}
 	}
 
-	if len(stats) != 5 {
-		t.Fatalf("Five stats must have been created, but was: %v", len(stats))
+	if len(stats) != 8 {
+		t.Fatalf("Eight stats must have been created, but was: %v", len(stats))
 	}
 
-	if stats[0].Path != "/" || stats[0].Visitors != 2 || stats[0].Browser.String != BrowserChrome || stats[0].BrowserVersion.String != "84.0" ||
-		stats[1].Path != "/page" || stats[1].Visitors != 1 || stats[1].Browser.String != BrowserChrome || stats[1].BrowserVersion.String != "84.0" ||
-		stats[2].Path != "/" || stats[2].Visitors != 1 || stats[2].Browser.String != BrowserFirefox || stats[2].BrowserVersion.String != "53.0" ||
-		stats[3].Path != "/" || stats[3].Visitors != 1 || stats[3].Browser.String != BrowserFirefox || stats[3].BrowserVersion.String != "54.0" ||
-		stats[4].Path != "/different-page" || stats[4].Visitors != 1 || stats[4].Browser.String != BrowserChrome || stats[4].BrowserVersion.String != "84.0" {
+	if stats[0].Path.String != "/" || stats[0].Visitors != 2 || stats[0].Browser.String != BrowserChrome || stats[0].BrowserVersion.String != "84.0" ||
+		stats[1].Path.String != "/page" || stats[1].Visitors != 1 || stats[1].Browser.String != BrowserChrome || stats[1].BrowserVersion.String != "84.0" ||
+		stats[2].Path.Valid || stats[2].Visitors != 3 || stats[2].Browser.String != BrowserChrome || stats[2].BrowserVersion.Valid ||
+		stats[3].Path.String != "/" || stats[3].Visitors != 1 || stats[3].Browser.String != BrowserFirefox || stats[3].BrowserVersion.String != "53.0" ||
+		stats[4].Path.String != "/" || stats[4].Visitors != 1 || stats[4].Browser.String != BrowserFirefox || stats[4].BrowserVersion.String != "54.0" ||
+		stats[5].Path.String != "/different-page" || stats[5].Visitors != 1 || stats[5].Browser.String != BrowserChrome || stats[5].BrowserVersion.String != "84.0" ||
+		stats[6].Path.Valid || stats[6].Visitors != 1 || stats[6].Browser.String != BrowserChrome || stats[6].BrowserVersion.Valid ||
+		stats[7].Path.Valid || stats[7].Visitors != 2 || stats[7].Browser.String != BrowserFirefox || stats[7].BrowserVersion.Valid {
 		t.Fatalf("Stats not as expected: %v", stats)
 	}
 }
@@ -359,4 +420,14 @@ func createHit(t *testing.T, store Store, tenantID int64, fingerprint, path, lan
 
 func day(year, month, day, hour int) time.Time {
 	return time.Date(year, time.Month(month), day, hour, 0, 0, 0, time.UTC)
+}
+
+func sumUpVisitors(stats []Stats) int {
+	sum := 0
+
+	for _, s := range stats {
+		sum += s.Visitors
+	}
+
+	return sum
 }